+++ apps/server/WorldObjects/WorldObject_Weapon.cs
@@ -1392,9 +1392,9 @@ partial class WorldObject
 
                 var scarabReduction = spell.School == MagicSchool.LifeMagic
                     ? playerAttacker.GetSigilTrinketManaReductionMod(spell, Skill.LifeMagic,
-                        (int)SigilTrinketLifeWarMagicEffect.Reduction)
+                        SigilTrinketLifeWarMagicEffect.Reduction)
                     : playerAttacker.GetSigilTrinketManaReductionMod(spell, Skill.WarMagic,
-                        (int)SigilTrinketLifeWarMagicEffect.Reduction);
+                        SigilTrinketLifeWarMagicEffect.Reduction);
 
                 if (playerAttacker.Mana.Current < (uint)(baseCost * scarabReduction))
                 {
